Add Python project definition and update README
The definition of the Python project is captured in a pyproject.toml
file, describing the build system, project information, and project
dependencies. The optinal dependencies required for testing only are
captured in the `dev` group.

The README is updated with instructions for setting up the local virtual
environment, the installation of IKOB therewithin, and basic testing
instructions.