Extract running of "stappen" into utility routine
This extracts logic from `cmdRun` of the ikobrunner into a separate
utility routine to enable reuse through end-to-end tests. The utility
implements the iteration over the subsequent steps of IKOB and
communicates its progress by yielding the completed step and
corresponding return code, which can be processed accordingly at the
corresponding call sites.