test: Introduce end-to-end testing for vlaanderen
This implements a basic end to end test for ikob, taking commit 110b99d
("Add end-to-end test", 2022-11-29) as an initial reference
implementation.

For a given project configuration, the end-to-end test will run through
all steps of the ikob process using the isolated iteration utility of
the previous commit. The test assert successful completion of each step
and will perform a complete directory comparison on the resulted output
files with respect to the stored reference files.

The Excel files are compared using their Pandas representations through
`pandas.testing.assert_frame_equal` considering the default settings for
absolute and relative tolerances on floating point comparisons.