feat: Add DataSource to Potentiebedrijven
NOTE: The value `mot` is used for formatting file paths through the
DataSource class, however, `mot` is just picked up from the last value
of the previous iteration. This does not seem right. For now, we make
this explicit by setting `mot` to a specific value paired with a
corresponding TODO marker.

NOTE: This odd use of `mot` is fixed in a later commit on this branch
once the DataSoure is propagated to all steps of the scripts.