feat: Remove keyword argument 'soort="matrix"'
The csvwegschrijven routine is able to determine the kind of data it is
writing by inferring if the provided data is an iterable of iterables.
If that is the case, the routine proceeds using `writer.writerows`,
otherwise `writer.writerow`.

Thereby it is no longer required to pass through the data kind into the
write_csv routine as keyword argument.