feat: Replace custom csv read/write with numpy
The custom routines for reading/writing lists of lists to/from csv files
are replaced by numpy loadtxt/savetxt routines. For writing the lists of
lists are first converted to numpy arrays, while for reading the
obtained numpy array is converted into lists of lists to maintain
compatible with the remainder of the code.

Up till now, the CSV files were compared directly, without introducing
additional relative/absolute tolerance in their comparisons. However, as
this patch modifies the read/write routines to numpy functions, small
floating point differences are observed with respect to the reference
files.

To address these small numerical differences, the CVS files are compared
using their pandas dataframe representation too, similar to the approach
that was already present for comparing the Excel files.