feat: Remove tracking if file has header or not
Previously, the DataSource class would keep track of which data files
had a header and informed the csv reader of how many lines to skip. This
ended up either being 1 (when a header was present), or 0 (when no
header was present).

Since these are the only options---either a header, or no header---the
adminstration can be avoided by trying to load the file to csv assuming
there is no header. If a header was present, this initial attempt will
fail due to the failing to convert the encountered string header to
float/int types. In that case, the loading is tried again, now assuming
that a single line header is indeed present.