fix: Patch filenames SEGS 'Verdeling_over_Groepen'
When introducing another test case (Eindhoven with 500 zones) a
regression was detected in writing of `Verdeling_over_groepen*` output
files, due to a filename postfix being left out.

This remained undetected in the original test case, since these files
were written back to the _input_ SEGS directory and therefore marked as
input files in the end-to-end test program.

Since the input directory of the second test case was not yet populated
with these files (which are actually output files) the regression became
clear as wrong files were written, and the to be expected SEGS files
could not be read from the input directory.

The regression is fixed and addressed in this patch by:

1. Moving the generated files previously marked as input to their
   expected location in the reference directory;
2. Restoring the postfix on the filename to ensure the expected files
   are written to disk.
3. Patching `DataSource` function `write_segs_*` to ensure that SEGS
   files being written are explicitly written to the temporary output
   directory and are not being mixed with input files.

A temporary patch is included in `DataSource` too that ensures reading
back previously written SEGS files (these exclusively start with
`Verdeling over groepen`) are diverted to the temporary output file.
Once this data is communicated in memory, this temporary patch will be
removed.