refactor: Extract file equality test for general use
Before `is_equal_file` was tailored to the directory comparison, where
it accepted the dircmp instance and the filename to compare. This
assumed the files were already marked as differing by the dircmp
comparison and we were only interested in reevaluating possible
differences between Excel/CSV files due to numerical tolerances.

This refactors `is_equal_file` to accept two files, a result and a
reference, and compares these files for equality. For Excel/CSV files
the dedicated comparison is performed, while for other files filecmp is
leveraged.