feat: Load Parkeerzoektijd or generate on the fly
Although the parkeerzoektijden can be directly generated from the
stedelijkheidsgraad, the parkeerzoektijden were explicitly stored on
disk.

This patch makes setting the parkeerzoektijden optional in the
configuration by:

1. Falling back to the expected parkeerzoektijden file when the
   specified file does not exist; or
2. By generating the parkeerzoektijden when the parkeerzoektijden file
   is either not set or not encountered on disk.

When both the parkeerzoektijden and stedelijkheidsgraad are missing an
assertion is raised, as we cannot recover from both files missing.

Corresponding tests are included to verify the expected
parkeerzoektijden are return for a couple different configuration
situations.