feat: Add missing loop autobezitgroepen arbeidsplaatsen
There are two similar functions calculating concurrentie, where the one
considering arbeid was missing the outer loop on autobezitgroepen. This
loop was also missing in the data generation during Potentiebedrijven.

This adds the missing loop and shifts the reference output files
accordingly, such that the "Herkomsten" data is located at the expected
location.