feat: Add DataKey class to get/set DataSource data
This defines the `DataKey` (data)class to group the required set of
strings to identify different data sources within the DataSource. The
dataclass effectively bundles the large number of parameters from which
the corresponding file paths are extracted.

This prepares for subsequent commits where these DataKeys can be
leveraged to keep the data cached in memory and passed around between
functions and thereby avoid the necessity to read/write data from disk
throughout the library.