fix: Do not reuse index when no minimum is found
The implementation to find the minimum position of the array, in this
specific case the minimum along `h` for `a` given `(i, j)` slice,
contains a bug where the previous index `hbewaar` is reused whenever all
values of that slice are above the initial minimum value.

For the current test problem this situation does appear, as there are
slices present in the data for which all values exceed `9999`. In those
cases the `hbewaar` value will be used as the position of the minimum
value, while this no longer corresponds to the position of the minimum
value in the array.

This patch removes the manual loop and relies on numpy argmin to
determine the position of the minimum value.

Since this fixes the minimum positions, the reference files need to be
updated accordingly.

Note: It seems worrisome that `9999` seems to be used as the largest
expected data, while the data clearly shows values exceeding this limit.
This observation was discussed with Hans and it was decided to continue
with this fix.