feat: Add ArgumentParser to control verbosity
A minimal argument parser definition is added to control verbosity of
the `ikobconfig` and `ikobrunner` GUIs. By default these scripts will
not relay the messages to `stdin`. Note, the helper scripts,
`ikobconfig.bat` and `ikobrunner.bat` invoke the GUIs with verbosity
enabled to provide some sense of progress while the application is
running.