fix: Avoid division by zero
Before the division by zero was allowed to happen and attempted to
recover from afterwards, i.e. after the division `np.where` was applied
to reset `inf` values to zero. However, it is also possible for the
division to generate `nan` values when zero is divided by zero. Those
cases were not covered by the division.

Also, since the interpreter will report the division by zero regardless
if we recover from it afterwards, it seems a better approach to simply
avoid the possible division by zero.