test: Add test case for divide by zero issue
This includes a test case to illustrate the divide by zero issue
observed in a specific test problem. It turns out this configuration is
different from the included end-to-end tests as it encounters not only
`x / 0.0` for `x` not equal to zero, but also `0.0 / 0.0` divisions.

This latter variant produces `nan` values rather than `inf` and since
only `inf` values were mapped to zero after the division, the resulting
arrays still contained `nan` values, which fail at point of writing to
disk.

The added test should be resolved by either recovering from the
generated `nan` values or by rewriting the division to prevent `inf` and
`nan` values to be created.