feat: Attempt to recover from known config changes
When changes are introduced to the GUI, e.g. by moving configuration
items to other tabs, the corresponding configuration files are altered
too. This causes old configuration files to be mark as invalid inputs,
as these are missing the new GUI definitions and therefore do not match
with what the validator expects.

Since the changes in GUI definitions are well defined, as typically only
some keys are moved within the configuration dictionaries, we can try to
recover from such invalid inputs by shifting these dictionary keys
accordingly. By doing so, we can maintain backward compatibility with
older configuration files.

Any attempt that is made to resolve such invalid configuration files are
logged to notify the user. To update old configuration files, they can
be loaded with `ikobconfig` and saved again. This should update the
configuration accordingly.