fix: Regression due to changes in baseline IKOB
The testing process by Hans indicated relatively large differences in
the output files like `Pot_totaal_Auto_OV_Fiets.xlsx`. Originally, this
was assumed to be a regression somewhere in the development history of
this version of IKOB. However, it turned out that identical results were
obtained in the *original* baseline version that was provided for the
consolidation, i.e. the version added in commit
`c44a15c1a049a638cc91d0ecd7d8cbb847deb512`.

Then, by comparing our baseline version of IKOB to the version that Hans
used to generate the reference files, it became clear that differences
were present in these two original implementations. After going through
these differences, two changes are responsible for the differences in
outputs observed by the testing of Hans:

* First, the alternative implementation contains a number of
  `np.transpose` calls that were not present in the original baseline.
  These are now propagated to the consolidated source code.
* Second, the alternative implementation does not use `group_list_1` in
  the second pass when extracting matrices. However, it overwite values
  in `group_list`. Again, this change was not present in the baseline
  version.

By transferring both changes onto the consolidated source code, we do
observe similar results as the references provided by Hans.

In subsequent communications with Hans it turned out that the second
change, the regression regarding `group_list` and `group_list_1` use,
was a regression in itself. Thus, this change does not need to be
implemented and only the transpose calls are introduced here.

The reference files are updated to reflect these changes.