feat: Make in memory caching configurable
The `DataSource` used to store all data within an in memory cache. It
turns out that for large simulation runs the cache might exceed the
system's RAM and the program crashes.

This patch makes this aspect of the `DataSource` configurable such that
users can opt in/out of this caching behaviour. Typically the program
will run faster with caching *enabled* due to less communication with
the file system, while running with caching *disabled* can lesses the
memory footprint of the application.